Feat: move TermConfig.vue
