+++ frontend/src/widgets/instance/dialogs/TermConfig.vue
